+++ packages/react-native-reanimated/src/css/native/normalization/animation/index.ts
@@ -3,3 +3,4 @@ export { normalizeAnimationKeyframes } from './keyframes';
 export { createSingleCSSAnimationProperties } from './properties';
 export { getAnimationSettingsUpdates } from './settings';
 export { normalizeSingleCSSAnimationSettings } from './settings';
+export { getAnimationsStartingStyle } from './startingStyle';